removed shell command
